removed ssh, apply.sh delete.sh more reliable

+++ bin/destroy.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 set -ex
 for d in application2 application1 shared network admin; do
-  if [ -d $d ]; then
+  if [ -d $d ] && [ -e $d/local.env ]; then
     (cd $d && source ./local.env && terraform destroy -auto-approve && rm local.env)
   fi
 done
